refactor: create files with fields configs
